Bug fix: Do not prepend `self.host` for `self.get` or `self.post`
While adding new features, I discovered that some of the existing APIs
for manipulating tokens and leases contained a bug where `self.get` and
related functions were prefixing `self.host` to URLs, but some of the
callers of `self.get` still thought they needed to prepend `self.host`
themselves.

So I fixed this issue everywhere that I could find it, and I added new
doc tests for as many of the affected APIs as possible.

The tests for two APIs are marked as `no_run`:

- `revoke` should not be run, because doing so would revoke our test
  token.
- `renew_lease` is currently disabled because the `get_secret` API
  currently offers no way to get a list ID for an existing secret.