API BREAKING: Handle missing `accessor` in `Auth` (Vault < 0.6.1)
This field is not provided by earlier versions of Vault.  We can simply
make it optional and it should work fine.

As far as I can tell, this was actually broken before, but I don't think
that any of our pre-existing test cases actually hit it.  So by adding
new tests for various APIs that we fixed, we revealed underlying
problems with an older version of vault.