Update to newer reqwest
The version of reqwest that vault-rs was using
depends on an older  version of OpenSSL that is
older than the version provided in Ubuntu Bionic+.
Because of this, vault-rs can no longer be built
on newer Ubuntu versions so we should update to a
version of reqwest that does support the newer
OpenSSL versions.

Updating reqwest requires several changes to how we
add headers to requests, as well as how the HTTP
methods are identified.

This change also bumps the travis test version to
use bionic rather than Trusty.